Added a simplified dropdown as `Select`
