+++ assets/src/edit-story/components/form/index.js
@@ -26,3 +26,4 @@ export { default as AdvancedDropDown } from './advancedDropDown';
 export { default as DateTime } from './dateTime';
 export { default as Required } from './required';
 export { default as RadioGroup } from './radioGroup';
+export { default as Select } from './select';
